--- src/components/common/index.ts
@@ -1,2 +1,0 @@
-export { default as TextField } from './TextField/TextField';
-export { default as Spacer } from './Spacer/Spacer';
