+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as TextField } from './TextField/TextField';
+export { default as Spacer } from './Spacer/Spacer';
+export { default as Input } from './Input/Input';
+export type { InputProps } from './Input/Input';
