+++ src/pages/payments/index.ts
@@ -1,0 +1,8 @@
+export { default as Payments } from './Payments/Payments';
+export { default as PaymentsPreview } from './payments-preview/PaymentsPreview/PaymentsPreview';
+export { default as CardNumberInput } from './payments-form/card-number/CardNumberInput/CardNumberInput';
+export { default as CardNumberTextField } from './payments-form/card-number/CardNumberTextField/CardNumberTextField';
+export { default as ExpirationDateInput } from './payments-form/expiration-date/ExpirationDateInput/ExpirationDateInput';
+export { default as ExpirationDateTextField } from './payments-form/expiration-date/ExpirationDateTextField/ExpirationDateTextField';
+export { default as OwnerNameInput } from './payments-form/owner-name/OwnerNameInput/OwnerNameInput';
+export { default as OwnerNameTextField } from './payments-form/owner-name/OwnerNameTextField/OwnerNameTextField';
