+++ vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
       { find: '@types', replacement: resolve(__dirname, 'src/types') },
       { find: '@utils', replacement: resolve(__dirname, 'src/utils') },
       { find: '@hooks', replacement: resolve(__dirname, 'src/hooks') },
+      { find: '@pages', replacement: resolve(__dirname, 'src/pages') },
     ],
   },
 });
